Split FreeInkUI.h into core + per-widget component headers
The monolithic 3371-line FreeInkUI.h is now a thin umbrella that includes
FreeInkUICore.h (geometry, draw target, interaction/Frame, generic helpers)
plus one header per widget under components/<group>/. Each widget header
colocates its props struct with its function and includes only its own
dependencies, so #include <FreeInkUI.h> consumers are unaffected.
